% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush-package.R
\docType{package}
\encoding{UTF-8}
\name{resmush-package}
\alias{resmush}
\alias{resmush-package}
\title{resmush: Optimize and Compress Image Files with 'reSmush.it'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Optimize and compress local and online image files with the 'reSmush.it' 'API' \url{https://resmush.it/api/}. Process individual files or entire directories. The 'API' is free for personal use, accepts files smaller than 5 'MB' and supports 'PNG', 'JPEG', 'GIF', 'BMP' and 'TIFF' files.
}
\seealso{
\itemize{
\item \code{\link[=resmush_file]{resmush_file()}} optimizes local image files.
\item \code{\link[=resmush_dir]{resmush_dir()}} optimizes image files in directories.
\item \code{\link[=resmush_url]{resmush_url()}} optimizes online image files.
\item \code{\link[=resmush_clean_dir]{resmush_clean_dir()}} removes suffixed output files.
\item The \href{https://dieghernan.github.io/resmush/}{package website} provides
articles and reference documentation.
\item The \href{https://github.com/dieghernan/resmush}{source code} is available on
GitHub.
\item \href{https://github.com/dieghernan/resmush/issues}{Report bugs} in the issue
tracker.
}
}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
