% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush-package.R
\docType{package}
\encoding{UTF-8}
\name{resmush-package}
\alias{resmush}
\alias{resmush-package}
\title{resmush: Optimize and Compress Image Files with 'reSmush.it'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Optimize and compress local image files, directories and online images with the 'reSmush.it' API \url{https://resmush.it/}. Supports 'png', 'jpg/jpeg', 'gif', 'bmp' and 'tiff' files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/resmush/}
  \item \url{https://github.com/dieghernan/resmush}
  \item Report bugs at \url{https://github.com/dieghernan/resmush/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
