% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush-package.R
\docType{package}
\name{resmush-package}
\alias{resmush}
\alias{resmush-package}
\title{resmush: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/resmush/}
  \item \url{https://github.com/dieghernan/resmush}
  \item Report bugs at \url{https://github.com/dieghernan/resmush/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
