% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush-clean-dir.R
\encoding{UTF-8}
\name{resmush_clean_dir}
\alias{resmush_clean_dir}
\title{Remove output files from directories}
\usage{
resmush_clean_dir(dir, suffix = "_resmush", recursive = FALSE)
}
\arguments{
\item{dir}{A character vector of directory paths. See the \code{path} argument of
\code{\link[base:list.files]{base::list.files()}}.}

\item{suffix}{A character string containing the suffix pattern used to
identify files. The value is interpreted as a regular expression. The
default is \code{"_resmush"}, which is also used by \code{\link[=resmush_file]{resmush_file()}}.}

\item{recursive}{Logical. Should the file search recurse into directories?}
}
\value{
An \code{\link[base:invisible]{base::invisible()}} \code{NULL}. Messages list the files selected for removal.
}
\description{
\strong{Use with caution.} Remove files that match \code{suffix} from one or more
directories. This function cleans output files created by
\code{\link[=resmush_file]{resmush_file()}} or \code{\link[=resmush_dir]{resmush_dir()}}.
}
\examples{
\donttest{
# Create a temporary file with a suffix to remove.
png_file <- system.file("extimg/example.png", package = "resmush")
suffix <- "_would_be_removed"
tmp_png <- file.path(
  tempdir(),
  paste0("example", suffix, ".png")
)

file.exists(tmp_png)
file.copy(png_file, tmp_png, overwrite = TRUE)

file.exists(tmp_png)

# Run with the default suffix. This should not remove the file.
resmush_clean_dir(tempdir())

file.exists(tmp_png)

# Use the matching suffix to remove the file.
resmush_clean_dir(tempdir(), suffix = suffix)

file.exists(tmp_png)
}
}
\seealso{
\code{\link[=resmush_file]{resmush_file()}} and \code{\link[=resmush_dir]{resmush_dir()}} create the suffixed output
files that this function removes.
}
\concept{helpers}
