% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_clean_dir.R
\encoding{UTF-8}
\name{resmush_clean_dir}
\alias{resmush_clean_dir}
\title{Clean files created by \CRANpkg{resmush}}
\usage{
resmush_clean_dir(dir, suffix = "_resmush", recursive = FALSE)
}
\arguments{
\item{dir}{A character vector of full path names. See the \code{path} argument in
\code{\link[=list.files]{list.files()}}.}

\item{suffix}{Character. Defaults to \code{"_resmush"}. See \code{\link[=resmush_file]{resmush_file()}}.}

\item{recursive}{Logical. Should the file search recurse into directories?}
}
\value{
Returns an \code{\link[=invisible]{invisible()}} \code{NULL} value and produces messages summarizing the
process.
}
\description{
\strong{Use with caution.} This removes files from your computer.

Clean a directory (or a list of directories) of files created by
\code{\link[=resmush_file]{resmush_file()}}.
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
\donttest{
# Simple example.

png_file <- system.file("extimg/example.png", package = "resmush")

# Copy to a temporary file with a given suffix.
suffix <- "_would_be_removed"
tmp_png <- file.path(
  tempdir(),
  paste0("example", suffix, ".png")
)

file.exists(tmp_png)
file.copy(png_file, tmp_png, overwrite = TRUE)

file.exists(tmp_png)

# Run with the default suffix; this should not remove the file.
resmush_clean_dir(tempdir())

file.exists(tmp_png)

# Use the matching suffix to remove the file.
resmush_clean_dir(tempdir(), suffix = suffix)

file.exists(tmp_png)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=resmush_file]{resmush_file()}}, \code{\link[=resmush_dir]{resmush_dir()}}, \code{\link[=list.files]{list.files()}}, \code{\link[=unlink]{unlink()}}.
}
\concept{helpers}
\keyword{internal}
