% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_dir.R
\encoding{UTF-8}
\name{resmush_dir}
\alias{resmush_dir}
\title{Optimize image files in directories}
\usage{
resmush_dir(
  dir,
  ext = "\\\\.(png|jpe?g|bmp|gif|tif)$",
  suffix = "_resmush",
  overwrite = FALSE,
  progress = TRUE,
  report = TRUE,
  recursive = FALSE,
  ...
)
}
\arguments{
\item{dir}{A character vector of paths to local directories.}

\item{ext}{A \code{\link[base:regex]{regex}} indicating the extensions of the files
to optimize. The default captures all extensions supported by the API.}

\item{suffix}{Character. Defaults to \code{"_resmush"}. By default, optimized
files are saved in the same directory with this suffix. For example,
\code{example.png} becomes \code{example_resmush.png}. Values \code{""}, \code{NA} and \code{NULL}
are equivalent to \code{overwrite = TRUE}.}

\item{overwrite}{Logical. Should the files in \code{dir} be overwritten? If \code{TRUE}
\code{suffix} is ignored.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{report}{Logical. Should a summary report be displayed in the console?}

\item{recursive}{Logical. Should file search within \code{dir} be recursive? See
also \code{\link[=list.files]{list.files()}}.}

\item{...}{
  Arguments passed on to \code{\link{resmush_file}}
  \describe{
    \item{\code{qlty}}{Integer between \code{0} and \code{100} indicating the optimization level.
This only affects \code{jpg} files. For optimal results, use values above \code{90}.}
    \item{\code{exif_preserve}}{Logical. Should
\href{https://en.wikipedia.org/wiki/Exif}{Exif} metadata be preserved? The
default is \code{FALSE}, which removes it.}
  }}
}
\value{
Writes optimized files to disk when the API call is successful. Invisibly
returns a data frame summarizing the process.
}
\description{
Optimize supported image files in one or more directories with the
\href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
\donttest{
# Copy the example directory.
example_dir <- system.file("extimg", package = "resmush")
temp_dir <- tempdir()
file.copy(example_dir, temp_dir, recursive = TRUE)

# Create the destination folder path.
dest_folder <- file.path(tempdir(), "extimg")

# Non-recursive.
resmush_dir(dest_folder)
resmush_clean_dir(dest_folder)

# Recursive.
summary <- resmush_dir(dest_folder, recursive = TRUE)

# Return the same information in the invisible data frame.
summary[, -c(1, 2)]

# Display the `png` output.
if (require("png", quietly = TRUE)) {
  a_png <- grepl("png$", summary$dest_img)
  my_png <- png::readPNG(summary[a_png, ]$dest_img[2])
  grid::grid.raster(my_png)
}

# Clean up the example files.
unlink(dest_folder, force = TRUE, recursive = TRUE)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api/}{reSmush.it API} documentation.

See \code{\link[=resmush_clean_dir]{resmush_clean_dir()}} to clean a directory of previous runs.

Other functions for optimizing:
\code{\link[=resmush_file]{resmush_file()}},
\code{\link[=resmush_url]{resmush_url()}}
}
\concept{optimize}
