% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_dir.R
\name{resmush_dir}
\alias{resmush_dir}
\title{Optimize files of several directories}
\usage{
resmush_dir(
  dir,
  ext = "\\\\.(png|jpe?g|bmp|gif|tif|webp)$",
  suffix = "_resmush",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dir}{Character or vector of characters representing paths of local
directories.}

\item{ext}{\link{regex} indicating the extensions of the files to be optimized.
The default value would capture all the extensions admitted by the API.}

\item{suffix}{Character, defaults to \code{"_resmush"}. By default, a new file
with the suffix is created in the same directory (i.e.,
optimized \code{example.png} would be \code{example_resmush.png}). Use \code{""}, \code{NA}
of \code{NULL} for overwrite the files.}

\item{verbose}{Logical. If \code{TRUE} displays a summary of the results.}

\item{...}{
  Arguments passed on to \code{\link[=resmush_file]{resmush_file}}
  \describe{
    \item{\code{qlty}}{Only affects \code{jpg} files. Integer between 0 and 100 indicating
the optimization level. For optimal results use vales above 90.}
    \item{\code{exif_preserve}}{Logical. Should the
\href{https://en.wikipedia.org/wiki/Exif}{Exif} information (if any) deleted?
Default is to remove (i.e. \code{exif_preserve = FALSE}).}
  }}
}
\value{
Writes on disk the optimized file if the API call is successful in the
directories specified in \code{dir}.
In any case, a (invisibly) data frame with a summary of the process is
returned as well.
}
\description{
Optimize all the local files of a directory (or list of directories) using
the  \href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
# Get example files and copy to a temporary dir
example_files <- list.files(system.file(package = "resmush"),
  pattern = "\\\\.(png|jpe?g)$",
  recursive = TRUE,
  full.names = TRUE
)

temp_dir <- file.path(tempdir(), "example_dir")
if (!dir.exists(temp_dir)) dir.create(temp_dir)
file.copy(example_files, temp_dir)

resmush_dir(temp_dir, verbose = TRUE)

# Clean up
unlink(temp_dir, force = TRUE)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api}{reSmush.it API} docs.

See \code{\link[=resmush_clean_dir]{resmush_clean_dir()}} to clean a directory of previous runs.

Other functions for optimizing:
\code{\link{resmush_file}()},
\code{\link{resmush_url}()}
}
\concept{optimize}
