% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_file.R
\encoding{UTF-8}
\name{resmush_file}
\alias{resmush_file}
\title{Optimize local image files}
\usage{
resmush_file(
  file,
  suffix = "_resmush",
  overwrite = FALSE,
  progress = TRUE,
  report = TRUE,
  qlty = 92,
  exif_preserve = FALSE
)
}
\arguments{
\item{file}{Path or paths to local image files. The API can optimize
\code{png}, \code{jpg/jpeg}, \code{gif}, \code{bmp} and \code{tiff} files.}

\item{suffix}{Character. Defaults to \code{"_resmush"}. By default, the optimized
file is saved in the same directory as \code{file} with this suffix. For
example, \code{example.png} becomes \code{example_resmush.png}. Values \code{""}, \code{NA} and
\code{NULL} are equivalent to \code{overwrite = TRUE}.}

\item{overwrite}{Logical. Should the file in \code{file} be overwritten? If \code{TRUE}
\code{suffix} is ignored.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{report}{Logical. Should a summary report be displayed in the console?}

\item{qlty}{Integer between \code{0} and \code{100} indicating the optimization level.
This only affects \code{jpg} files. For optimal results, use values above \code{90}.}

\item{exif_preserve}{Logical. Should
\href{https://en.wikipedia.org/wiki/Exif}{Exif} metadata be preserved? The
default is \code{FALSE}, which removes it.}
}
\value{
Writes optimized files to disk when the API call is successful. Invisibly
returns a data frame summarizing the process. With \code{report = TRUE}, a summary
report is also displayed in the console.
}
\description{
Optimize one or more local image files with the
\href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}

\donttest{
png_file <- system.file("extimg/example.png", package = "resmush")

# Copy to a temporary file for this example.
tmp_png <- tempfile(fileext = ".png")

file.copy(png_file, tmp_png, overwrite = TRUE)

resmush_file(tmp_png)

# Several paths.
jpg_file <- system.file("extimg/example.jpg", package = "resmush")
tmp_jpg <- tempfile(fileext = ".jpg")

file.copy(jpg_file, tmp_jpg, overwrite = TRUE)

# Display a summary in the console.
summary <- resmush_file(c(tmp_png, tmp_jpg))

# The invisible data frame contains the same information.
summary

# Display the `png` output.
if (require("png", quietly = TRUE)) {
  my_png <- png::readPNG(summary$dest_img[1])
  grid::grid.raster(my_png)
}

# Use with `jpg` files and parameters.
resmush_file(tmp_jpg)
resmush_file(tmp_jpg, qlty = 10)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api/}{reSmush.it API} documentation.

See \code{\link[=resmush_clean_dir]{resmush_clean_dir()}} to clean a directory of previous runs.

Other functions for optimizing:
\code{\link[=resmush_dir]{resmush_dir()}},
\code{\link[=resmush_url]{resmush_url()}}
}
\concept{optimize}
