% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_file.R
\name{resmush_file}
\alias{resmush_file}
\title{Optimize a local file}
\usage{
resmush_file(
  file,
  suffix = "_resmush",
  qlty = 92,
  verbose = FALSE,
  exif_preserve = FALSE
)
}
\arguments{
\item{file}{Path or paths to local files. \strong{reSmush} can optimize the
following image files:
\itemize{
\item \code{png}
\item \code{jpg/jpeg}
\item \code{gif}
\item \code{bmp}
\item \code{tiff}
\item \code{webp}
}}

\item{suffix}{Character, defaults to \code{"_resmush"}. By default, a new file
with the suffix is created in the same directory than \code{file}. (i.e.,
optimized \code{example.png} would be \code{example_resmush.png}). Use \code{""}, \code{NA}
of \code{NULL} for overwrite the initial \code{file}.
the \code{file} use any of the following values: \verb{"", NA, NULL}.}

\item{qlty}{Only affects \code{jpg} files. Integer between 0 and 100 indicating
the optimization level. For optimal results use vales above 90.}

\item{verbose}{Logical. If \code{TRUE} displays a summary of the results.}

\item{exif_preserve}{Logical. Should the
\href{https://en.wikipedia.org/wiki/Exif}{Exif} information (if any) deleted?
Default is to remove (i.e. \code{exif_preserve = FALSE}).}
}
\value{
Writes on disk the optimized file if the API call is successful in the
same directory than \code{file}.
In any case, a (invisibly) data frame with a summary of the process is
returned as well.
}
\description{
Optimize local images using the \href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

\donttest{
png_file <- system.file("extimg/example.png", package = "resmush")

# For the example, copy to a temporary file
tmp_png <- tempfile(fileext = ".png")

file.copy(png_file, tmp_png, overwrite = TRUE)

resmush_file(tmp_png)

# Several paths
jpg_file <- system.file("extimg/example.jpg", package = "resmush")
tmp_jpg <- tempfile(fileext = ".jpg")

file.copy(jpg_file, tmp_jpg, overwrite = TRUE)

summary <- resmush_file(c(tmp_png, tmp_jpg))

# Returns an (invisible) data frame with a summary of the process
summary


# With parameters

# Silently returns a data frame
resmush_file(tmp_jpg, verbose = TRUE)
resmush_file(tmp_jpg, verbose = TRUE, qlty = 10)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api}{reSmush.it API} docs.

See \code{\link[=resmush_clean_dir]{resmush_clean_dir()}} to clean a directory of previous runs.

Other functions for optimizing:
\code{\link{resmush_dir}()},
\code{\link{resmush_url}()}
}
\concept{optimize}
