% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_file.R
\name{resmush_file}
\alias{resmush_file}
\title{Optimize a local file}
\usage{
resmush_file(
  file,
  outfile = file,
  qlty = 92,
  verbose = FALSE,
  exif_preserve = FALSE
)
}
\arguments{
\item{file}{Path or paths to a local file. \strong{reSmush} can optimize the
following image files:
\itemize{
\item \code{png}
\item \code{jpg}
\item \code{gif}
\item \code{bmp}
\item \code{tif}
}}

\item{outfile}{Path or paths where the optimized file would be store in
your disk. By default, it would override the file specified in \code{file}. It
should be of the same length than \code{file} parameter.}

\item{qlty}{Only affects \code{jpg} files. Integer between 0 and 100 indicating
the optimization level. For optimal results use vales above 90.}

\item{verbose}{Logical. If \code{TRUE} displays a summary of the results.}

\item{exif_preserve}{Logical. Should be the
\href{https://en.wikipedia.org/wiki/Exif}{Exif} information removed as well?
Default is to remove (i.e. \code{exif_preserve = FALSE}).}
}
\value{
Writes on disk the optimized file if the API call is successful.
In any case, a (invisibly) data frame with a summary of the process is
returned as well.
}
\description{
Optimize local images using the \href{https://resmush.it/}{reSmush.it API}.

\strong{Note that} the default parameters of the function \code{outfile = file}
\strong{overwrites} the local file. See \strong{Examples}.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

\donttest{
png_file <- system.file("extimg/example.png", package = "resmush")

# For the example, write to a temp file
tmp_png <- tempfile(fileext = ".png")

resmush_file(png_file, outfile = tmp_png)

# Several paths
jpg_file <- system.file("extimg/example.jpg", package = "resmush")
tmp_jpg <- tempfile(fileext = ".jpg")


summary <- resmush_file(c(png_file, jpg_file), outfile = c(
  tmp_png,
  tmp_jpg
))

# Returns an (invisible) data frame with a summary of the process
summary


# With parameters

# Silently returns a data frame
resmush_file(jpg_file, outfile = tmp_jpg, verbose = TRUE)
resmush_file(jpg_file, outfile = tmp_jpg, verbose = TRUE, qlty = 10)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api}{reSmush.it API} docs.

\code{\link[=resmush_url]{resmush_url()}}
}
