% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_url.R
\encoding{UTF-8}
\name{resmush_url}
\alias{resmush_url}
\title{Optimize online image files}
\usage{
resmush_url(
  url,
  outfile = file.path(tempdir(), basename(url)),
  overwrite = FALSE,
  progress = TRUE,
  report = TRUE,
  qlty = 92,
  exif_preserve = FALSE
)
}
\arguments{
\item{url}{A character vector of URLs pointing to hosted image files. The API
can optimize PNG, JPEG, GIF, BMP and TIFF files.}

\item{outfile}{A character vector of paths where optimized files are stored.
By default, files are created in \code{\link[=tempdir]{tempdir()}} with the same \code{\link[=basename]{basename()}} as
each file in \code{url}. \code{outfile} must have the same length as \code{url}.}

\item{overwrite}{Logical. Should \code{outfile} be overwritten if it already
exists? If \code{FALSE} and \code{outfile} exists, a copy is created with a
numeric suffix, such as \verb{<outfile>_01.png}.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{report}{Logical. Should a summary report be displayed in the console?}

\item{qlty}{An integer between \code{0} and \code{100} indicating the optimization
level. This only affects JPEG files. For optimal results, use values above
\code{90}.}

\item{exif_preserve}{Logical. Should
\href{https://en.wikipedia.org/wiki/Exif}{Exif} metadata be preserved? The
default is \code{FALSE}, which removes it.}
}
\value{
A data frame summarizing the optimization, returned invisibly. Successful
API calls also write the optimized files to disk. If \code{outfile} contains
duplicate paths, \code{resmush_url()} makes them unique with suffixes such as
\verb{_01} and \verb{_02}.
}
\description{
Optimize and download one or more online image files with the
\href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}

\donttest{
# Base URL.
base_url <- "https://raw.githubusercontent.com/dieghernan/resmush/main/inst/"

png_url <- paste0(base_url, "/extimg/example.png")
resmush_url(png_url)

# Several URLs.
jpg_url <- paste0(base_url, "/extimg/example.jpg")

summary <- resmush_url(c(png_url, jpg_url))

# The invisible data frame contains a summary of the process.
summary

# Display the PNG output.
if (require("png", quietly = TRUE)) {
  my_png <- png::readPNG(summary$dest_img[1])
  grid::grid.raster(my_png)
}

# Use with JPEG files and parameters.
resmush_url(jpg_url)
resmush_url(jpg_url, qlty = 10)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api/}{reSmush.it API} documentation.

Other image optimization functions:
\code{\link[=resmush_dir]{resmush_dir()}},
\code{\link[=resmush_file]{resmush_file()}}
}
\concept{optimize}
