% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_url.R
\encoding{UTF-8}
\name{resmush_url}
\alias{resmush_url}
\title{Optimize online image files}
\usage{
resmush_url(
  url,
  outfile = file.path(tempdir(), basename(url)),
  overwrite = FALSE,
  progress = TRUE,
  report = TRUE,
  qlty = 92,
  exif_preserve = FALSE
)
}
\arguments{
\item{url}{URL or vector of URLs pointing to hosted image files. The API can
optimize \code{png}, \code{jpg/jpeg}, \code{gif}, \code{bmp} and \code{tiff} files.}

\item{outfile}{Path or paths where optimized files are stored on disk. By
default, temporary files are created with \code{\link[=tempfile]{tempfile()}} and the same
\code{\link[=basename]{basename()}} as the file provided in \code{url}. \code{outfile} must have the same
length as \code{url}.}

\item{overwrite}{Logical. Should \code{outfile} be overwritten if it already
exists? If \code{FALSE} and \code{outfile} exists, a copy is created with a
numerical suffix, such as \verb{<outfile>_01.png}.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{report}{Logical. Should a summary report be displayed in the console?}

\item{qlty}{Integer between \code{0} and \code{100} indicating the optimization level.
This only affects \code{jpg} files. For optimal results, use values above \code{90}.}

\item{exif_preserve}{Logical. Should
\href{https://en.wikipedia.org/wiki/Exif}{Exif} metadata be preserved? The
default is \code{FALSE}, which removes it.}
}
\value{
Writes optimized files to disk when the API call is successful. Invisibly
returns a data frame summarizing the process. If any value in \code{outfile} is
duplicated, \code{resmush_url()} renames the outputs with suffixes such as \verb{_01}
and \verb{_02}.
}
\description{
Optimize and download one or more online image files with the
\href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}

\donttest{
# Base URL.
base_url <- "https://raw.githubusercontent.com/dieghernan/resmush/main/inst/"

png_url <- paste0(base_url, "/extimg/example.png")
resmush_url(png_url)

# Several URLs.
jpg_url <- paste0(base_url, "/extimg/example.jpg")

summary <- resmush_url(c(png_url, jpg_url))

# The invisible data frame contains a summary of the process.
summary

# Display the `png` output.
if (require("png", quietly = TRUE)) {
  my_png <- png::readPNG(summary$dest_img[1])
  grid::grid.raster(my_png)
}

# Use with `jpg` files and parameters.
resmush_url(jpg_url)
resmush_url(jpg_url, qlty = 10)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api/}{reSmush.it API} documentation.

Other functions for optimizing:
\code{\link[=resmush_dir]{resmush_dir()}},
\code{\link[=resmush_file]{resmush_file()}}
}
\concept{optimize}
