% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmush_url.R
\name{resmush_url}
\alias{resmush_url}
\title{Optimize an online file}
\usage{
resmush_url(
  url,
  outfile = file.path(tempdir(), basename(url)),
  qlty = 92,
  verbose = FALSE
)
}
\arguments{
\item{url}{url to a hosted file. \strong{reSmush} can optimize the
following image files:
\itemize{
\item \code{png}
\item \code{jpg}
\item \code{gif}
\item \code{bmp}
\item \code{tif}
}}

\item{outfile}{Path where the optimized file would be store in your disk. By
default, a temporary file (see \code{\link[=tempfile]{tempfile()}}) with the same \code{\link[=basename]{basename()}} than
the file provided in url would be created.}

\item{qlty}{Only affects \code{jpg} files. Integer between 0 and 100 indicating
the optimization level. For optimal results use vales above 90.}

\item{verbose}{Logical. If \code{TRUE} displays a summary of the results.}
}
\value{
Writes on disk the optimized file if the API call is successful.
In any case, a (invisibly) data frame with a summary of the process is
returned as well.
}
\description{
Optimize and download an online image using the
\href{https://resmush.it/}{reSmush.it API}.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

\donttest{

# Base url
base_url <- "https://raw.githubusercontent.com/dieghernan/resmush/main/inst/"

png_url <- paste0(base_url, "/extimg/example.png")

# Silently returns a data frame
png_res <- resmush_url(png_url)
png_res

# Use with jpg and parameters
jpg_url <- paste0(base_url, "/extimg/example.jpg")

# Silently returns a data frame
resmush_url(jpg_url, verbose = TRUE)
resmush_url(jpg_url, verbose = TRUE, qlty = 10)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://resmush.it/api}{reSmush.it API} docs.

\code{\link[=resmush_file]{resmush_file()}}
}
